+++ Classes/GRMustacheVersion.h
@@ -33,12 +33,12 @@
  
  @since v1.0
  */
-#define GRMUSTACHE_MINOR_VERSION 8
+#define GRMUSTACHE_MINOR_VERSION 9
 
 /**
  The patch-level component of GRMustache version
  
  @since v1.0
  */
-#define GRMUSTACHE_PATCH_VERSION 6
+#define GRMUSTACHE_PATCH_VERSION 0