+++ include/GRMustache.h
@@ -34,7 +34,7 @@ typedef NSUInteger GRMustacheTemplateOptions;
 /**
  A C struct that hold GRMustache version information
  
- @since v1.0.0
+ @since v1.0
  */
 typedef struct {
 	int major;	/**< The major component of the version. */
@@ -46,7 +46,7 @@ typedef struct {
 /**
  The GRMustache class provides with global-level information and configuration
  of the GRMustache library.
- @since v1.0.0
+ @since v1.0
  */
 @interface GRMustache: NSObject
 
@@ -55,22 +55,20 @@ typedef struct {
 //////////////////////////////////////////////////////////////////////////////////////////
 
 /**
- @returns the version of GRMustache as a GRMustacheVersion.
+ Returns the version of GRMustache as a GRMustacheVersion struct.
  
- @since v1.0.0
+ @return The version of GRMustache as a GRMustacheVersion struct.
+ @since v1.0
  */
 + (GRMustacheVersion)version AVAILABLE_GRMUSTACHE_VERSION_1_0_AND_LATER;
 
 //////////////////////////////////////////////////////////////////////////////////////////
-/// @name Handling of boolean properties
+/// @name Handling of BOOL properties
 //////////////////////////////////////////////////////////////////////////////////////////
 
 /**
- A Boolean value that determines whether GRMustache renders templates in strict
+ Returns a Boolean value that determines whether GRMustache renders templates in strict
  boolean mode.
- 
- @returns YES if GRMustache renders templates in strict boolean mode,
- NO otherwise. The default value is NO.
  
  In strict boolean mode, properties of context objects that are declared as BOOL
  are interpreted as numbers, and can not be used for controlling Mustache
@@ -80,28 +78,80 @@ typedef struct {
  those declared as BOOL), are interpreted as booleans, and can be used for
  controlling Mustache boolean sections.
  
- @see GRMustache#setStrictBooleanMode:
- @see GRYes
- @see GRNo
- @since v1.0.0
+ @return YES if GRMustache renders templates in strict boolean mode,
+ NO otherwise. The default value is NO.
+ 
+ @see setStrictBooleanMode:
+ @see [GRMustacheContext valueForKey:]
+ @since v1.0
  */
 + (BOOL)strictBooleanMode AVAILABLE_GRMUSTACHE_VERSION_1_0_AND_LATER;
 
 /**
  Sets the strict boolean mode of GMustache.
  
- @param aBool YES if GRMustache should render templates in strict boolean mode,
+ See the documentation for strictBooleanMode for more information.
+ 
+ @param strictBooleanMode YES if GRMustache should render templates in strict boolean mode,
  NO otherwise.
  
- @see GRMustache#strictBooleanMode
- @since v1.0.0
+ @see strictBooleanMode
+ @see [GRMustacheContext valueForKey:]
+ @since v1.0
  */
-+ (void)setStrictBooleanMode:(BOOL)aBool AVAILABLE_GRMUSTACHE_VERSION_1_0_AND_LATER;
++ (void)setStrictBooleanMode:(BOOL)strictBooleanMode AVAILABLE_GRMUSTACHE_VERSION_1_0_AND_LATER;
 
+
+//////////////////////////////////////////////////////////////////////////////////////////
+/// @name Preventing NSUndefinedKeyException when using GRMustache in Development configuration
+//////////////////////////////////////////////////////////////////////////////////////////
+
+/**
+ Have GRMustache raise much less `NSUndefinedKeyExceptions` when rendering templates.
+ 
+ The rendering of a GRMustache template can lead to many `NSUndefinedKeyExceptions` to be raised, because of the heavy usage of Key-Value Coding. Those exceptions are nicely handled by GRMustache, and are part of the regular rendering of a template.
+ 
+ Unfortunately, when debugging a project, developers usually set their debugger to stop on every Objective-C exceptions. GRMustache rendering can thus become a huge annoyance. This method prevents it.
+ 
+ You'll get a slight performance hit, so you'd probably make sure this call does not enter your Release configuration.
+ 
+ One way to achieve this is to add `-DDEBUG` to the "Other C Flags" setting of your development configuration, and to wrap the `preventNSUndefinedKeyExceptionAttack` method call in a #if block, like:
+ 
+    #ifdef DEBUG
+    [GRMustache preventNSUndefinedKeyExceptionAttack];
+    #endif
+ 
+ @since v1.7
+ */
 + (void)preventNSUndefinedKeyExceptionAttack AVAILABLE_GRMUSTACHE_VERSION_1_7_AND_LATER;
 
+
+//////////////////////////////////////////////////////////////////////////////////////////
+/// @name Global template options
+//////////////////////////////////////////////////////////////////////////////////////////
+
+/**
+ Returns the default template options. The default value is GRMustacheTemplateOptionNone.
+ 
+ @return The default template options. The default value is GRMustacheTemplateOptionNone.
+ @see setDefaultTemplateOptions:
+ @since v1.8
+ */
 + (GRMustacheTemplateOptions)defaultTemplateOptions AVAILABLE_GRMUSTACHE_VERSION_1_8_AND_LATER;
 
+/**
+ Sets the default template options.
+ 
+ Those options will be used by all GRMustacheTemplate rendering and parsing methods, such as [GRMustacheTemplate parseString:error:] and [GRMustacheTemplate renderObject:fromString:error:].
+ 
+ For instance, you'll trigger support for the [Mustache Specification 1.1.2](https://github.com/mustache/spec) with:
+ 
+    [GRMustache setDefaultTemplateOptions:GRMustacheTemplateOptionMustacheSpecCompatibility];
+ 
+ @param templateOptions A mask of options indicating the default behavior of templates.
+ @see defaultTemplateOptions
+ @since v1.8
+ */
 + (void)setDefaultTemplateOptions:(GRMustacheTemplateOptions)templateOptions AVAILABLE_GRMUSTACHE_VERSION_1_8_AND_LATER;
 
 @end